+++ quanttide_docs/cli/tutorial.py
@@ -11,13 +11,30 @@ from quanttide_docs.models.tutorial import Tutorial
 
 
 @cli.command()
+def validate(path='.'):
+    """
+    验证教程格式
+
+    :param path:本地仓库路径，默认为当前目录
+    :return:
+    """
+    with Tutorial(local_path=path) as tutorial:
+        # 基础信息
+        typer.echo(f"课程名称：{tutorial.name}")
+        if hasattr(tutorial, 'depot_name') and tutorial.depot_name:
+            typer.echo(f"课程仓库名称：{tutorial.depot_name}")
+        if hasattr(tutorial, 'local_path') and tutorial.local_path:
+            typer.echo(f"本地仓库位置：{tutorial.local_path}")
+        # 解析
+        return tutorial.is_valid()
+
+
+@cli.command()
 def preview(path='.'):
     """
     预览当前提交的解析结果。
 
-    :param name: 课程名称，必填。
     :param path: 本地仓库路径，默认为当前目录
-    :param result_path: 解析结果存储路径，默认为_build目录的qtclass.json
     :return:
     """
     # 构建兼验证